:truck: move SseListener to tests.util
